add: ファイル名の変更+レーダー項目が変わっても動的に対応できるように変更
